Load and Capacity Test of an Web application using JMeter
The performance of the Address Book application will be examined after increasing the volume of data in the database. In our project we will be performing various CRUD commands on the database. CRUD means create, read, update, and delete commands will be used to add any record, read record from database, update any existing record, or delete the existing record.
To perform the load and volume testing, we used JMeter to create a test plan with two thread group:
1.	Thread group for web load testing
2.	Thread group for Volume testing